experiment with lifecycle and onDestroy() method of main Activity
the idea is that in the onDestroy() method of the activity, it should persist the contents of shopping cart. This method is called when the activity is destroyed. If this doesn't work. Consider using the same code on onPause().
Furthermore, to not block the app closing down, I'm trying to use lifecycle scope and coroutines