feat: database created, connected to a repository and viewModel
It seems that is working correctly. The viewModel has correct connection to the repository. And the injection of dependencies seems to be working alright.

I'm not really sure how the dependency injection worked or why. But It worked and the database is in the file system. Furthermore It stays there when the app is closed. So I'm happy with how things turned out :D